Extract modal analytical solver and tighten checks
Extract modal analytical solver and tighten checks
